+++ internal/database/entity/comment.go
@@ -3,10 +3,10 @@ package entity
 import "time"
 
 type Comment struct {
-	ID        uint `gorm:"primaryKey;autoIncrement"`
+	ID        uint64 `gorm:"primaryKey;autoIncrement"`
 	Comment   string
-	UserID    uint
-	ContentID uint
+	UserID    uint64
+	ContentID uint64
 	CreatedAt time.Time `gorm:"autoCreateTime"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime"`
 }
