+++ internal/database/entity/feedback.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type Feedback struct {
-	ID        uint `gorm:"primaryKey;autoIncrement"`
-	UserID    uint
+	ID        uint64 `gorm:"primaryKey;autoIncrement"`
+	UserID    uint64
 	Message   string
 	CreatedAt time.Time `gorm:"autoCreateTime"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime"`
