+++ internal/database/entity/follow.go
@@ -3,9 +3,9 @@ package entity
 import "time"
 
 type Follow struct {
-	ID          uint `gorm:"primaryKey;autoIncrement"`
-	FollowerID  uint
-	FollowingID uint
+	ID          uint64 `gorm:"primaryKey;autoIncrement"`
+	FollowerID  uint64
+	FollowingID uint64
 	CreatedAt   time.Time `gorm:"autoCreateTime"`
 	UpdatedAt   time.Time `gorm:"autoUpdateTime"`
 }
