+++ internal/database/entity/likeContent.go
@@ -3,9 +3,9 @@ package entity
 import "time"
 
 type LikeContent struct {
-	ID        uint `gorm:"primaryKey;autoIncrement"`
-	UserID    uint
-	ContentID uint
+	ID        uint64 `gorm:"primaryKey;autoIncrement"`
+	UserID    uint64
+	ContentID uint64
 	CreatedAt time.Time `gorm:"autoCreateTime"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime"`
 }
