+++ internal/database/entity/notification.go
@@ -3,11 +3,11 @@ package entity
 import "time"
 
 type Notification struct {
-	ID         uint `gorm:"primaryKey;autoIncrement"`
+	ID         uint64 `gorm:"primaryKey;autoIncrement"`
 	Content    string
 	IsChecked  bool `gorm:"default:false"`
-	ReceiverID uint
-	TriggerID  uint
+	ReceiverID uint64
+	TriggerID  uint64
 	CreatedAt  time.Time `gorm:"autoCreateTime"`
 	UpdatedAt  time.Time `gorm:"autoUpdateTime"`
 }
