+++ internal/modules/aIConversation/handler.go
@@ -10,8 +10,8 @@ type Handler struct {
 }
 
 type input struct {
-	ID           uint `json:"id"`
-	UserID       uint `json:"userId"`
+	ID           uint64 `json:"id"`
+	UserID       uint64 `json:"userId"`
 }
 
 func NewHandler(aIConversationService AIConversationService) Handler {
