+++ internal/modules/aIMessage/repository.go
@@ -6,11 +6,11 @@ import (
 )
 
 type AIMessageRepository interface {
-	Create(senderId uint, conversationId uint, message string, response string) (*entity.AIMessage, error)
-	FindAll(conversationId uint) (*[]entity.AIMessage, error)
-	FindOne(id uint) (*entity.AIMessage, error)
-	Update(id uint, message string) (*entity.AIMessage, error)
-	DeleteOne(id uint) error
+	Create(senderId uint64, conversationId uint64, message string, response string) (*entity.AIMessage, error)
+	FindAll(conversationId uint64) (*[]entity.AIMessage, error)
+	FindOne(id uint64) (*entity.AIMessage, error)
+	Update(id uint64, message string) (*entity.AIMessage, error)
+	DeleteOne(id uint64) error
 }
 
 type aIMessageRepository struct {
@@ -21,7 +21,7 @@ func NewAIMessageRepository(db *gorm.DB) AIMessageRepository {
 	return &aIMessageRepository{db: db}
 }
 
-func (r *aIMessageRepository) Create(senderId uint, conversationId uint, message string, response string) (*entity.AIMessage, error) {
+func (r *aIMessageRepository) Create(senderId uint64, conversationId uint64, message string, response string) (*entity.AIMessage, error) {
 	aIMessage := entity.AIMessage{
 		SenderID: senderId,
 		ConversationID: conversationId,
@@ -38,7 +38,7 @@ func (r *aIMessageRepository) Create(senderId uint, conversationId uint, message
 	return &aIMessage, nil
 }
 
-func (r *aIMessageRepository) FindAll(conversationId uint) (*[]entity.AIMessage, error) {
+func (r *aIMessageRepository) FindAll(conversationId uint64) (*[]entity.AIMessage, error) {
 	var aiMessages []entity.AIMessage
 
 	err := r.db.Where("conversation_id = ?", conversationId).Find(&aiMessages).Error
@@ -50,7 +50,7 @@ func (r *aIMessageRepository) FindAll(conversationId uint) (*[]entity.AIMessage,
 	return &aiMessages, nil
 }
 
-func (r *aIMessageRepository) FindOne(id uint) (*entity.AIMessage, error) {
+func (r *aIMessageRepository) FindOne(id uint64) (*entity.AIMessage, error) {
 	var aIMessage entity.AIMessage
 
 	err := r.db.Where("id = ?", id).Take(&aIMessage).Error
@@ -62,7 +62,7 @@ func (r *aIMessageRepository) FindOne(id uint) (*entity.AIMessage, error) {
 	return &aIMessage, nil
 }
 
-func (r *aIMessageRepository) Update(id uint, message string) (*entity.AIMessage, error) {
+func (r *aIMessageRepository) Update(id uint64, message string) (*entity.AIMessage, error) {
 	aIMessage, err := r.FindOne(id)
 
 	if err != nil {
@@ -75,7 +75,7 @@ func (r *aIMessageRepository) Update(id uint, message string) (*entity.AIMessage
 	return aIMessage, nil
 }
 
-func (r *aIMessageRepository) DeleteOne(id uint) error {
+func (r *aIMessageRepository) DeleteOne(id uint64) error {
 
 	err := r.db.Delete(&entity.AIMessage{}, id).Error
