+++ internal/modules/comment/repository.go
@@ -6,11 +6,11 @@ import (
 )
 
 type CommentRepository interface {
-	Create(userId uint, contentId uint, text string) (*entity.Comment, error)
-	FindAll(contentId uint) (*[]entity.Comment, error)
-	FindOne(id uint) (*entity.Comment, error)
-	Update(id uint, text string) (*entity.Comment, error)
-	DeleteOne(id uint) error
+	Create(userId uint64, contentId uint64, text string) (*entity.Comment, error)
+	FindAll(contentId uint64) (*[]entity.Comment, error)
+	FindOne(id uint64) (*entity.Comment, error)
+	Update(id uint64, text string) (*entity.Comment, error)
+	DeleteOne(id uint64) error
 }
 
 type commentRepository struct {
@@ -21,7 +21,7 @@ func NewContentRepository(db *gorm.DB) CommentRepository {
 	return &commentRepository{db: db}
 }
 
-func (r *commentRepository) Create(userId uint, contentId uint, text string) (*entity.Comment, error) {
+func (r *commentRepository) Create(userId uint64, contentId uint64, text string) (*entity.Comment, error) {
 	comment := entity.Comment{
 		UserID:    userId,
 		ContentID: contentId,
@@ -36,7 +36,7 @@ func (r *commentRepository) Create(userId uint, contentId uint, text string) (*e
 	return &comment, nil
 }
 
-func (r *commentRepository) FindAll(contentId uint) (*[]entity.Comment, error) {
+func (r *commentRepository) FindAll(contentId uint64) (*[]entity.Comment, error) {
 	var comments []entity.Comment
 
 	err := r.db.Where("content_id = ?", contentId).Find(&comments).Error
@@ -48,7 +48,7 @@ func (r *commentRepository) FindAll(contentId uint) (*[]entity.Comment, error) {
 	return &comments, nil
 }
 
-func (r *commentRepository) FindOne(id uint) (*entity.Comment, error) {
+func (r *commentRepository) FindOne(id uint64) (*entity.Comment, error) {
 	var comment entity.Comment
 
 	err := r.db.Where("id = ?", id).Take(&comment).Error
@@ -60,7 +60,7 @@ func (r *commentRepository) FindOne(id uint) (*entity.Comment, error) {
 	return &comment, nil
 }
 
-func (r *commentRepository) Update(id uint, text string) (*entity.Comment, error) {
+func (r *commentRepository) Update(id uint64, text string) (*entity.Comment, error) {
 	comment, err := r.FindOne(id)
 
 	if err != nil {
@@ -73,7 +73,7 @@ func (r *commentRepository) Update(id uint, text string) (*entity.Comment, error
 	return comment, nil
 }
 
-func (r *commentRepository) DeleteOne(id uint) error {
+func (r *commentRepository) DeleteOne(id uint64) error {
 
 	err := r.db.Delete(&entity.Comment{}, id).Error
