+++ internal/modules/groupChat/handler.go
@@ -12,8 +12,8 @@ type Handler struct {
 }
 
 type input struct {
-	ID         uint `json:"id"`
-	UserID     uint `json:"userId"`
+	ID         uint64 `json:"id"`
+	UserID     uint64 `json:"userId"`
 	Name       string `json:"name"`
 	Members    []entity.User
 	PictureUrl string `json:"pictureUrl"`
